RT and response time
I added the RT and the response time. I also tried out if the experiment worked and it didn't at first, until I changed the pygame.event.clear(pygame.KEYDOWN) to pygame.event.clear.  Otherwise it would do the first round properly and the rest not so well. I also changed the PCT_NOGO to 0.2, as it is calculated times the NUMTRIAL in the main